(#587) Implement scripts_of_svg.py script
