+++ .gitignore
@@ -2,3 +2,4 @@ build/
 GPATH
 GRTAGS
 GTAGS
+__pycache__/
