+++ devtools/svg/tags.py
@@ -1,0 +1,2 @@
+RECT_TAG_NAME = '{http://www.w3.org/2000/svg}rect'
+TEXT_TAG_NAME = '{http://www.w3.org/2000/svg}text'
